requirements.txt
gym==0.26.0
torch>=1.10.0
numpy>=1.20.0
pyyaml
